feat: add ratelimiter task interval config.

+++ polaris-common/polaris-config/src/main/java/com/tencent/polaris/api/config/provider/RateLimitConfig.java
@@ -83,7 +83,7 @@ public interface RateLimitConfig extends PluginConfig, Verifier {
 
     long getRemoteTaskIntervalMilli();
 
-    long getRangeDelayMilli();
+    int getRangeDelayMilli();
     /**
      * 是否上报限流监控视图, 该开关默认关闭，如果需要使用限流的老监控视图，则需要开启此监控数据上报开关
      *
